+++ src/Symfony/Component/DependencyInjection/Extension/ConfigurableExtensionInterface.php
@@ -23,10 +23,10 @@ interface ConfigurableExtensionInterface extends ConfigurableInterface
     /**
      * Allows an extension to prepend the extension configurations.
      */
-    public function prependExtension(ContainerConfigurator $container, ContainerBuilder $builder): void;
+    public function prependExtension(ContainerConfigurator $configurator, ContainerBuilder $container): void;
 
     /**
      * Loads a specific configuration.
      */
-    public function loadExtension(array $config, ContainerConfigurator $container, ContainerBuilder $builder): void;
+    public function loadExtension(array $config, ContainerConfigurator $configurator, ContainerBuilder $container): void;
 }