+++ src/Symfony/Component/DependencyInjection/Kernel/AbstractBundle.php
@@ -54,11 +54,11 @@ abstract class AbstractBundle implements BundleInterface, ConfigurableExtensionI
     {
     }
 
-    public function prependExtension(ContainerConfigurator $container, ContainerBuilder $builder): void
+    public function prependExtension(ContainerConfigurator $configurator, ContainerBuilder $container): void
     {
     }
 
-    public function loadExtension(array $config, ContainerConfigurator $container, ContainerBuilder $builder): void
+    public function loadExtension(array $config, ContainerConfigurator $configurator, ContainerBuilder $container): void
     {
     }